+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+* Add offchain account support for the tonic client method `get_account_update`.
 * Refactorized `get_account` to create the account from a single query.
 * Admit partial account IDs for the commands that need them.
 * Added nextest to be used as test runner